requirements.txt
alembic==1.6.5
amqp==5.0.6
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.4.1
bcrypt==3.2.0
billiard==3.6.4.0
black==21.6b0
celery==5.1.2
cffi==1.14.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
fastapi==0.66.0
flower==1.0.0
graphene==2.1.8
graphene-sqlalchemy==2.3.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==1.1.0
h11==0.12.0
humanize==3.10.0
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
psycopg2-binary==2.9.1
pycparser==2.20
pydantic==1.8.2
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
redis==3.5.3
regex==2021.7.1
Rx==1.6.1
singledispatch==3.6.2
six==1.16.0
SQLAlchemy==1.4.20
starlette==0.14.2
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.0
uvicorn==0.14.0
vine==5.0.0
wcwidth==0.2.5

